make MetadataItem private
